requirements.txt
fastapi
uvicorn
python-jose
python-multipart
python-dotenv
